requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
yt-dlp
opencv-python
mediapipe
numpy
Pillow
torch
transformers
pyannote.audio
ultralytics
openai
waitress
faster-whisper
SQLAlchemy